more of program loops

+++ src/util/probability.tcc
@@ -51,7 +51,7 @@ random_draw_from_integer_pdf(const S& p) {
 template <class S>
 size_t
 random_draw_from_integer_cdf(const S& s) {
-	static const int64_t lrand_max = 1LL <<32;
+	static const int64_t lrand_max = int64_t(1) <<32;
 	const int64_t total = long(s.back());
 	const int64_t b_rem = lrand_max % total;
 	const int64_t upper = lrand_max - b_rem;
